+++ src/components/header/Navigate.tsx
@@ -12,8 +12,8 @@ import { useRouter } from 'next/router';
 import useAuth from '../../hooks/useAuth';
 import { getAuth, signOut } from 'firebase/auth';
 import app from '../../firebaseConfig';
-import Logout from './Logout12';
-import Login from './Login12';
+import Logout from './Logout';
+import Login from './Login';
 import { getCookie } from 'cookies-next';
 import { tokenName } from '../../helpers/const';
